make number of boot HDD configurable
This patch extends e543c1f3 to allow any boot HDD from A to C.
It unhardcodes the value "2" for boot HDD in MOSINIT2.ASM.
This change does not affect the floppy boot because of:

test al,80h ; hard drive?
jz doflpy   ; no

in boot code.
The number of floppies is saved into "sbcflops" var, and is
currently not used anywhere. This is for further extensions.
I had no success in booting from drives past C but this is
unlikely being of any importance.