MAKEMOS.BAT: improve PATH handling [fixes #30]
This patch makes the path of MAKEMOS.BAT to be appended
_after_ the original path, that is usually prepared by build.sh.
Also remove the BORLANDC from path, as it is not needed during
build and may fail because of an incompatible make.
Also dos5.0 and lib are removed, because, as the original
path is now preserved, the build can happen under any DOS -
it doesn't need to be self-hosted and use pre-built MOS binaries.
And finally build.sh is updated to the new dosemu2 syntax.
The old one seems to have been removed.