build: Tidy up the RSA includes and libs
1/ There were multiple identical copies of RSASMALL.LIB so delete the
duplicates and move it and RSALARGE.LIB to libs directory. Listed both
the large and small libraries to confirm that the contents were
consistent with each other and the header file.

2/ There were multiple similar copies of RSA.H so after determining that
the diffs were mainly whitespace, the following differences remained:
  a) commented out parameter.
    -int         restore_video_region(int, int, void *);
    +int         restore_video_region(int, int) /*, void *);*/
    It can be seen from the callers of it throughout the repo that the
    third parameter is used, so that version is the one taken.
  b) commented out 'FOREVER' macro
    Some users of RSA.H define this themselves, so use the commented
version. Other users may also have to do the same.